backend/pyproject.toml
[project]
name = "luxis-backend"
version = "0.1.0"
description = "Luxis — Practice Management System for Dutch Law Firms"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",

    # Settings
    "pydantic-settings>=2.7.0",

    # Task queue
    "celery[redis]>=5.4.0",

    # Document generation
    "docxtpl>=0.18.0",
    "weasyprint>=62.0",

    # File uploads
    "python-multipart>=0.0.9",

    # Email
    "aiosmtplib>=3.0.0",
    "jinja2>=3.1.0",

    # HTTP client (for OAuth token exchange and Gmail API)
    "httpx>=0.28.0",

    # Encryption (for storing OAuth tokens at rest)
    "cryptography>=44.0.0",

    # Date parsing (for email date headers)
    "python-dateutil>=2.9.0",

    # Scheduler
    "apscheduler>=3.10.0,<4.0.0",

    # Sentry (optional)
    "sentry-sdk[fastapi]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
